requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.1
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.2.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
django-debug-toolbar~=3.6.0
